% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/felix.R
\name{get_hierarchy}
\alias{get_hierarchy}
\title{Gets (or plots the distribution of) the top n most frequent species in an area.}
\usage{
get_hierarchy(areaData, ranking, plot = FALSE)
}
\arguments{
\item{areaData}{Data for a specific area.}

\item{ranking}{Number for the top n most common species in the area.}

\item{plot}{Whether to plot or just return the data.}
}
\value{
If `plot` is false, a n-row dataframe containing the top n most common species
        with columns containing species_id, frequency, and species_name is returned.
        If `plot` is true, a plotly pie plot of the population distribution based on the generated
        dataframe is returned.
}
\description{
Gets (or plots the distribution of) the top n most frequent species in an area.
}
