% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/larris.R
\name{get_region}
\alias{get_region}
\title{Generates a named list of the region data in a specific provided region}
\usage{
get_region(type, name)
}
\arguments{
\item{type}{The type of region}

\item{name}{The name of the regional code or area}
}
\value{
The named list containing the information of the specified region
}
\description{
Generates a named list of the region data in a specific provided region
}
