% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred.R
\name{get_species_id}
\alias{get_species_id}
\title{Searches for a species by a name, returning the first ID in the search results}
\usage{
get_species_id(species_name)
}
\arguments{
\item{species_name}{A name related to a species}
}
\value{
The ID of the first search result for `species_name`
}
\description{
Searches for a species by a name, returning the first ID in the search results
}
