% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred.R
\name{get_species_name}
\alias{get_species_name}
\title{Gets the English name for a species given its ID, falling back to the scientific name}
\usage{
get_species_name(id)
}
\arguments{
\item{id}{The ID of the species}
}
\value{
The English or scientific name of the given species
}
\description{
Gets the English name for a species given its ID, falling back to the scientific name
}
