% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred.R
\name{most_common_across_regions}
\alias{most_common_across_regions}
\title{Gets (or plots) the most common species in each area of the provided data}
\usage{
most_common_across_regions(species, granularity = 4, plot = F)
}
\arguments{
\item{species}{Any species data}

\item{granularity}{The side length of areas to analyze (in degrees)}

\item{plot}{Whether to plot or just return the data}
}
\value{
If `plot` is false, a list of info for each area; the info is a named 2-element list,
        containing `region`, which has the starting and ending longitude/latitude coordinates
        for that area (`c(lon_1, lat_1, lon_2, lat_2)`), and `max_species`, which is itself a
        named 2-element list containing the ID of the most common species in the area, and a
        tibble of all the sightings of that species in the area
        If `plot` is true, a plot of all the areas in the data which contain sightings, colored
        according to the most common species in the area
}
\description{
Gets (or plots) the most common species in each area of the provided data
}
