% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junbo.R
\name{observations_in_region}
\alias{observations_in_region}
\title{Finds observations in a region defined by a polygon}
\usage{
observations_in_region(region, dataset)
}
\arguments{
\item{region}{list with numeric matrices with points in rows}

\item{dataset}{data.frame of observations}
}
\value{
`dataset` of observations inside `region`
}
\description{
Finds observations in a region defined by a polygon
}
