% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/felix.R
\name{plot_density}
\alias{plot_density}
\title{Plots the population density of a species.}
\usage{
plot_density(speciesData)
}
\arguments{
\item{speciesData}{Data for a single species}
}
\value{
A plotly densitymapbox figure of the species location data.
}
\description{
Plots the population density of a species.
}
