% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/felix.R
\name{plot_path}
\alias{plot_path}
\title{Plots the estimated migration path of a species filtered by year.}
\usage{
plot_path(speciesData)
}
\arguments{
\item{speciesData}{Data for a single species}
}
\value{
A plotly mapbox figure of segments making up the path of a species with a togglable year filter.
}
\description{
Plots the estimated migration path of a species filtered by year.
}
