% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junbo.R
\name{plot_sight_per_prov}
\alias{plot_sight_per_prov}
\title{Plots the bar graph of the number of sightings per province}
\usage{
plot_sight_per_prov(dataset)
}
\arguments{
\item{dataset}{data.frame of observations}
}
\value{
Bar graph of the number of sightings per province for `dataset`
}
\description{
Plots the bar graph of the number of sightings per province
}
