% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junbo.R
\name{plot_sightings_per_year}
\alias{plot_sightings_per_year}
\title{Plots the line graph of the number of sightings per year}
\usage{
plot_sightings_per_year(dataset)
}
\arguments{
\item{dataset}{data.frame of observations}
}
\value{
Line graph of the number of sightings per year for `dataset`
}
\description{
Plots the line graph of the number of sightings per year
}
