% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred.R
\name{plot_species_limited}
\alias{plot_species_limited}
\title{Plots a limited number of sightings of the given data}
\usage{
plot_species_limited(species, limit_points = 500)
}
\arguments{
\item{species}{Species data with longitude and latitude info}

\item{limit_points}{The max number of sightings to plot}
}
\value{
A plot of at most `limit_points` sightings from `species`
}
\description{
Plots a limited number of sightings of the given data
}
