% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junbo.R
\name{relative_observation_trends}
\alias{relative_observation_trends}
\title{Calculates the relative observation trend for a species}
\usage{
relative_observation_trends(species_id, dataset)
}
\arguments{
\item{species_id}{species_id of the chosen species}

\item{dataset}{data.frame of observations}
}
\value{
relative observation trends for the `species_id` in the `dataset`
}
\description{
Calculates the relative observation trend for a species
}
