% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred.R
\name{species_max_location}
\alias{species_max_location}
\title{Gets (or plots) the area containing the most sightings of a species}
\usage{
species_max_location(species, granularity = 4, plot = F, limit_points = 500)
}
\arguments{
\item{species}{Data for a single species}

\item{granularity}{The side length of areas to check (in degrees)}

\item{plot}{Whether to plot or just return the data}

\item{limit_points}{The max number of points to plot (if `plot` is true)}
}
\value{
If `plot` is false, a 5-element vector with the starting and ending longitude/latitude
        coordinates for the area containing the most sightings of the species, as well as the
        number of sightings (`c(lon_1, lat_1, lon_2, lat_2, n)`)
        If `plot` is true, a plot of at most `limit_points` sightings and a highlighted box
        representing the area with the most sightings
}
\description{
Gets (or plots) the area containing the most sightings of a species
}
