% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred.R
\name{species_max_statprov}
\alias{species_max_statprov}
\title{Gets the province or state containing the most sightings of the species in the given data}
\usage{
species_max_statprov(species)
}
\arguments{
\item{species}{Data for a single species}
}
\value{
A 2-element vector containing the state/province code (i.e. "ON") and the number of
        sightings of the species in that area
}
\description{
Gets the province or state containing the most sightings of the species in the given data
}
